requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Cors==3.0.10
gunicorn==20.0.4
idna==3.4
importlib-metadata==5.2.0
itsdangerous==2.1.2
jellyfish==0.9.0
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.24.1
python-decouple==3.6
pytz==2022.7
requests==2.28.1
six==1.16.0
urllib3==1.26.13
Werkzeug==2.2.2
yat-geo-db==1.0.3
zipp==3.11.0
